pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "beautifulsoup4>=4.14.2",
    "celery>=5.5.3",
    "docker>=7.1.0",
    "eventlet>=0.40.3",
    "flask-limiter>=4.0.0",
    "flask-sock>=0.7.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "minio>=7.2.18",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.11",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pyyaml>=6.0.3",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
]
